*** empty log message ***
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk/M2@1083 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/e/mem.cpp
@@ -132,9 +132,9 @@ void stash::text_out(buffer &o) const
 	  name, 
 	  (element_size * highwater + 1023)/1024,
 	  (element_size * n_inuse + 1023)/1024,
-	  element_size,
-	  n_allocs,
-	  n_inuse,
+	  (int)element_size,
+	  (int)n_allocs,
+	  (int)n_inuse,
 	  highwater,
 	  n_frees,
 	  newline);
